requirements.txt
# Package dependencies
PyYAML>=5.1

# Linters, formatters, security checkers
black==19.3b0
flake8==3.7.7
isort==4.3.20
mypy==0.701
safety

# Testing
pytest==4.6.2
pytest-cov==2.7.1

# Development tools
ipython
ipdb
